third_party/WebKit/LayoutTests/svg/animations/animate-number-calcMode-discrete-keyTimes-expected.txt

CONSOLE WARNING: SVG's SMIL animations (<animate>, <set>, etc.) are deprecated and will be removed. Please use CSS animations or Web animations instead.
SVG 1.1 dynamic animation tests

Test calcMode discrete with from-to animation on numbers. You should see a green 100x100 rect and only PASS messages

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect.x.animVal.value is 100
PASS rect.x.baseVal.value is 100
PASS rect.x.animVal.value is 100
PASS rect.x.baseVal.value is 100
PASS rect.x.animVal.value is 200
PASS rect.x.baseVal.value is 100
PASS rect.x.animVal.value is 200
PASS rect.x.baseVal.value is 100
PASS rect.x.animVal.value is 300
PASS rect.x.baseVal.value is 100
PASS successfullyParsed is true

TEST COMPLETE
